+++ crates/astria-bridge-withdrawer/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update `idna` dependency to resolve cargo audit warning [#1869](https://github.com/astriaorg/astria/pull/1869).
+- Support FROST threshold signing using signer nodes. [#1948](https://github.com/astriaorg/astria/pull/1948).
 
 ## [1.0.1] - 2024-11-01